second try to get the topic multi-target to work in a maintenance update without disrupting already running sap cluster environments
