+++ SAPHana/SAPHanaSR-ScaleOut.changes
@@ -1,7 +1,7 @@
 -------------------------------------------------------------------
-Fri Aug 28 15:16:18 UTC 2020 - abriel@suse.com
+Thu Jan  7 14:30:44 UTC 2021 - abriel@suse.com
 
-- change version to 0.170.4 - Testversion
+- change version to 0.180.0 - Testversion
 - change promotion scoring during maintenance procedure to prevent
   that both sides have an equal promotion scoring after refresh
   which might result in a critical promotion of the secondary.
@@ -40,6 +40,28 @@ Fri Aug 28 15:16:18 UTC 2020 - abriel@suse.com
   SAPHanaSR-showAttr.8 (bsc#1144312)
 - adjust README file
   (bsc#1144729)
+
+-------------------------------------------------------------------
+Wed Nov  4 10:10:06 UTC 2020 - abriel@suse.com
+
+- change version to 0.164.1
+- adapt man page SAPHanaSR-showAttr(8) and the README
+  (bsc#1144729)
+
+-------------------------------------------------------------------
+Mon Sep 21 14:40:04 UTC 2020 - abriel@suse.com
+
+- PROMOTED/PROMOTED Fix - The PROMOTED/PROMOTED values happened
+  after the main tenance procedure and the refresh of the resource
+  did not fixed that
+  (bsc#1176330)
+- Improved SCORING and logging
+  Score of secondary in takeover phase increased from 122 to 145 to
+  avoid promotion of former primary masternameserver candidates
+  (bsc#1174610)
+- Fixed typos and improved descriptions in comments
+- Change default timeouts and intervals to match the official
+  recommendations
 
 -------------------------------------------------------------------
 Thu May 14 15:28:09 UTC 2020 - abriel@suse.com
