+++ SAPHana/man/SAPHanaSR-ScaleOut.7
@@ -1,4 +1,4 @@
-.\" Version: 0.170.0
+.\" Version: 0.180.0
 .\"
 .TH SAPHanaSR-ScaleOut 7 "03 Aug 2020" "" "SAPHanaSR-ScaleOut"
 .\"
