+++ SAPHana/man/SAPHanaSR-ScaleOut_basic_cluster.7
@@ -1,4 +1,4 @@
-.\" Version: 0.170.0
+.\" Version: 0.180.0
 .\"
 .TH SAPHanaSR-ScaleOut 7 "27 Jul 2020" "" "SAPHanaSR-ScaleOut_basic_cluster"
 .\"
