+++ SAPHana/man/SAPHanaSR-showAttr.8
@@ -1,4 +1,4 @@
-.\" Version: 0.170.0
+.\" Version: 0.180.0
 .\"
 .TH SAPHanaSR-showAttr 8 "12 Aug 2020" "" "SAPHanaSR-ScaleOut"
 .\"
