+++ SAPHana/man/SAPHanaSrMultiTarget.py.7
@@ -1,4 +1,4 @@
-.\" Version: 0.170.0
+.\" Version: 0.180.0
 .\"
 .TH SAPHanaSrMultiTarget.py 7 "13 Aug 2020" "" "SAPHanaSR-ScaleOut"
 .\"
