+++ SAPHana/man/ocf_suse_SAPHanaController.7
@@ -1,4 +1,4 @@
-.\" Version: 0.170.0
+.\" Version: 0.180.0
 .\"
 .TH ocf_suse_SAPHanaController 7 "11 Aug 2020" "" "OCF resource agents"
 .\"
