+++ SAPHana/srHook/SAPHanaSR.py
@@ -3,171 +3,135 @@
 # Author:       Fabian Herschel, 2015
 # License:      GNU General Public License (GPL)
 # Copyright:    (c) 2015-2016 SUSE Linux GmbH
-# Copyright:    (c) 2017-2020 SUSE LLC
-
-### ANGI TODO - check # SAPHanaSR needs SAP HANA 2.0 SPS4 (2.00.040.00) as minimum version
+# Copyright:    (c) 2017-2021 SUSE LLC
 """
-import os, time
-
-fhSRHookVersion = "0.170.5.1001.1647"
-
-try:
-    from hdb_ha_dr.client import HADRBase
-except ImportError as e:
-    print("Module HADRBase not found - running outside of SAP HANA? - {0}".format(e))
+from hdb_ha_dr.client import HADRBase
+import subprocess, os
 
 """
-Only for SAP HANA >= 2.0 SPS3
-
-To use this HA/DR hook provide please add the following lines (or similar) to your global.ini:
+Sample for a HA/DR hook provider.
+When using your own code in here, please copy this file to location on
+/hana/shared outside the HANA installation.
+To use this HA/DR hook provide please add the following lines (or
+similar) to your global.ini:
     [ha_dr_provider_SAPHanaSR]
     provider = SAPHanaSR
-    path = /usr/share/SAPHanaSR-ScaleOut
+    path = /hana/shared/myHooks/SAPHanaSR-ScaleOut
     execution_order = 1
 
     [trace]
     ha_dr_saphanasr = info
 """
+fhSRHookVersion = "0.180.0.0812.1853"
+srHookGen = "1.0"
 
-try:
-    class SAPHanaSR(HADRBase):
 
-        def __init__(self, *args, **kwargs):
-            # delegate construction to base class
-            super(SAPHanaSR, self).__init__(*args, **kwargs)
-            method = "init"
-            self.tracer.info("{0}.{1}() version {2}".format(self.__class__.__name__, method, fhSRHookVersion))
+class SAPHanaSR(HADRBase):
 
-        def about(self):
-            method = "about"
-            self.tracer.info("{0}.{1}() version {2}".format(self.__class__.__name__, method, fhSRHookVersion))
-            return {"provider_company": "SUSE",
-                    "provider_name": "SAPHanaSR",  # class name
-                    "provider_description": "Inform Cluster about SR state",
-                    "provider_version": "1.0"}
+    def __init__(self, *args, **kwargs):
+        # delegate construction to base class
+        super(SAPHanaSR, self).__init__(*args, **kwargs)
+        method = "init"
+        self.tracer.info("{0}.{1}() version {2}".format(self.__class__.__name__, method, fhSRHookVersion))
+        mySID = os.environ.get('SAPSYSTEMNAME')
+        mysid = mySID.lower()
+        myNode = subprocess.check_output(['sudo', '/usr/sbin/crm_node', '-n'])
+        myCMD = "sudo /usr/sbin/crm_attribute -N {0} -v {1} -n hana_{2}_node_srHook_version -l reboot".format(myNode, srHookGen, mysid)
+        rc = os.system(myCMD)
+        myMSG = "CALLING CRM: <{0}> rc={1}".format(myCMD, rc)
+        self.tracer.info("{0}.{1}() {2}\n".format(self.__class__.__name__, method, myMSG))
+        self.tracer.info("{0}.{1}() Running old srHookGeneration {2}, see attribute hana_{3}_node_srHook_version too\n".format(self.__class__.__name__, method, srHookGen, mysid)
 
-        def startup(self, hostname, storage_partition, sr_mode, **kwargs):
-            method = "startup"
-            self.tracer.debug("enter startup hook; {0}".format(locals()))
-            self.tracer.debug(self.config.toString())
-            self.tracer.info("leave startup hook")
+    def about(self):
+        return {"provider_company": "SUSE",
+                "provider_name": "SAPHanaSR",  # class name
+                "provider_description": "Inform Cluster about SR state",
+                "provider_version": "1.0"}
+
+    def startup(self, hostname, storage_partition, sr_mode, **kwargs):
+        self.tracer.debug("enter startup hook; %s" % locals())
+        self.tracer.debug(self.config.toString())
+        self.tracer.info("leave startup hook")
+        return 0
+
+    def shutdown(self, hostname, storage_partition, sr_mode, **kwargs):
+        self.tracer.debug("enter shutdown hook; %s" % locals())
+        self.tracer.debug(self.config.toString())
+        self.tracer.info("leave shutdown hook")
+        return 0
+
+    def failover(self, hostname, storage_partition, sr_mode, **kwargs):
+        self.tracer.debug("enter failover hook; %s" % locals())
+        self.tracer.debug(self.config.toString())
+        self.tracer.info("leave failover hook")
+        return 0
+
+    def stonith(self, failingHost, **kwargs):
+        self.tracer.debug("enter stonith hook; %s" % locals())
+        self.tracer.debug(self.config.toString())
+        # e.g. stonith of params["failed_host"]
+        # e-g- set vIP active
+        self.tracer.info("leave stonith hook")
+        return 0
+
+    def preTakeover(self, isForce, **kwargs):
+        """Pre takeover hook."""
+        self.tracer.info("%s.preTakeover method called with isForce=%s" % (self.__class__.__name__, isForce))
+        if not isForce:
+            # run pre takeover code
+            # run pre-check, return != 0 in case of error => will abort takeover
+            return 0
+        else:
+            # possible force-takeover only code
+            # usually nothing to do here
             return 0
 
-        def shutdown(self, hostname, storage_partition, sr_mode, **kwargs):
-            method = "shutdown"
-            self.tracer.debug("enter shutdown hook; {0}".format(locals()))
-            self.tracer.debug(self.config.toString())
-            self.tracer.info("leave shutdown hook")
+    def postTakeover(self, rc, **kwargs):
+        """Post takeover hook."""
+        self.tracer.info("%s.postTakeover method called with rc=%s" % (self.__class__.__name__, rc))
+        if rc == 0:
+            # normal takeover succeeded
+            return 0
+        elif rc == 1:
+            # waiting for force takeover
+            return 0
+        elif rc == 2:
+            # error, something went wrong
             return 0
 
-        def failover(self, hostname, storage_partition, sr_mode, **kwargs):
-            method = "failover"
-            self.tracer.debug("enter failover hook; {0}".format(locals()))
-            self.tracer.debug(self.config.toString())
-            self.tracer.info("leave failover hook")
-            return 0
-
-        def stonith(self, failingHost, **kwargs):
-            method = "stonith"
-            self.tracer.debug("enter stonith hook; {0}".format(locals()))
-            self.tracer.debug(self.config.toString())
-            # e.g. stonith of params["failed_host"]
-            # e-g- set vIP active
-            self.tracer.info("leave stonith hook")
-            return 0
-
-        def preTakeover(self, isForce, **kwargs):
-            """Pre takeover hook."""
-            method = "preTakeover"
-            self.tracer.info("{0}.{1}() method called with isForce={2}".format(self.__class__.__name__, method, isForce))
-            if not isForce:
-                # run pre takeover code
-                # run pre-check, return != 0 in case of error => will abort takeover
-                return 0
+    def srConnectionChanged(self, ParamDict, **kwargs):
+        """ finally we got the srConnection hook :) """
+        self.tracer.info("SAPHanaSR (%s) %s.srConnectionChanged method called with Dict=%s" % (fhSRHookVersion, self.__class__.__name__, ParamDict))
+        # myHostname = socket.gethostname()
+        # myDatebase = ParamDict["database"]
+        mySystemStatus = ParamDict["system_status"]
+        mySID = os.environ.get('SAPSYSTEMNAME')
+        mysid = mySID.lower()
+        myInSync = ParamDict["is_in_sync"]
+        myReason = ParamDict["reason"]
+        myNode = subprocess.check_output(['sudo', '/usr/sbin/crm_node', '-n'])
+        myCMD = "sudo /usr/sbin/crm_attribute -N {0} -v {1} -n hana_{2}_node_srHook_version -l reboot".format(myNode, srHookGen, mysid)
+        rc = os.system(myCMD)
+        myMSG = "CALLING CRM: <{0}> rc={1}".format(myCMD, rc)
+        self.tracer.info("{0}.{1}() {2}\n".format(self.__class__.__name__, method, myMSG))
+        self.tracer.info("{0}.{1}() Running old srHookGeneration {2}, see attribute hana_{3}_node_srHook_version too\n".format(self.__class__.__name__, method, srHookGen, mysid)
+        if (mySystemStatus == 15):
+            mySRS = "SOK"
+        else:
+            if (myInSync):
+                # ignoring the SFAIL, because we are still in sync
+                self.tracer.info("SAPHanaSR (%s) %s.srConnectionChanged ignoring bad SR status because of is_in_sync=True (reason=%s)" % (fhSRHookVersion, self.__class__.__name__, myReason))
+                mySRS = ""
             else:
-                # possible force-takeover only code
-                # usually nothing to do here
-                return 0
-
-        def postTakeover(self, rc, **kwargs):
-            """Post takeover hook."""
-            method = "postTakeover"
-            self.tracer.info("{0}.{1}() method called with rc={2}".format(self.__class__.__name__, method, rc))
-            if rc == 0:
-                # normal takeover succeeded
-                return 0
-            elif rc == 1:
-                # waiting for force takeover
-                return 0
-            elif rc == 2:
-                # error, something went wrong
-                return 0
-
-        def srConnectionChanged(self, ParamDict, **kwargs):
-            """ finally we got the srConnection hook :) """
-            method = "srConnectionChanged"
-            self.tracer.info("{0}.{1}() method called with Dict={2} (version {3})".format(self.__class__.__name__, method, ParamDict, fhSRHookVersion))
-            # myHostname = socket.gethostname()
-            # myDatebase = ParamDict["database"]
-            globHOOK = False
-            mySystemStatus = ParamDict["system_status"]
-            mySID = os.environ.get('SAPSYSTEMNAME')
-            mysid = mySID.lower()
-            myInSync = ParamDict["is_in_sync"]
-            myReason = ParamDict["reason"]
-            mySite = ParamDict["siteName"]
-            if mySystemStatus == 15:
-                mySRS = "SOK"
-            else:
-                if myInSync:
-                    # ignoring the SFAIL, because we are still in sync
-                    self.tracer.info("{0}.{1}() ignoring bad SR status because of is_in_sync=True (reason={2})".format(self.__class__.__name__, method, myReason))
-                    mySRS = ""
-                else:
-                    mySRS = "SFAIL"
-            if mySRS == "":
-                myMSG = "### Ignoring bad SR status because of is_in_sync=True ###"
-                self.tracer.info("{0}.{1}() {2}\n".format( self.__class__.__name__, method, myMSG ))
-            elif mySite == "":
-                myMSG = "### Ignoring bad SR status because of empty site name in call params ###"
-                self.tracer.info("{0}.{1}() was called with empty site name. Ignoring call.".format(self.__class__.__name__, method))
-            else:
-                myCMD = "sudo /usr/sbin/crm_attribute -G -n hana_%s_glob_srHook" % (mysid)
-                rc = os.system(myCMD)
-                if rc == 0:
-                    # found global Hook attribute, write both (old and new) attributes
-                    # for compatibility reasons
-                    globHOOK = True
-                    myCMD = "sudo /usr/sbin/crm_attribute -n hana_{0}_glob_srHook -v {1} -t crm_config -s SAPHanaSR".format(mysid, mySRS)
-                    rc = os.system(myCMD)
-                    myMSG = "CALLING CRM: <{0}> rc={1}".format(myCMD, rc)
-                    self.tracer.info("{0}.{1}() {2}\n".format(self.__class__.__name__, method, myMSG))
-                myCMD = "sudo /usr/sbin/crm_attribute -n hana_{0}_site_srHook_{1} -v {2} -t crm_config -s SAPHanaSR".format(mysid, mySite, mySRS)
-                rc = os.system(myCMD)
-                myMSG = "CALLING CRM: <{0}> rc={1}".format(myCMD, rc)
-                self.tracer.info("{0}.{1}() {2}\n".format(self.__class__.__name__, method, myMSG))
-                #
-                if rc != 0:
-                    #
-                    # FALLBACK
-                    # sending attribute to the cluster failed - using failback method and write status to a file - RA to pick-up the value during next SAPHanaController monitor operation
-                    #
-                    myMSG = "sending attribute to the cluster failed - using local file as fallback"
-                    self.tracer.info("{0}.{1}() {2}\n".format( self.__class__.__name__, method, myMSG ))
-                    #
-                    # cwd of hana is /hana/shared/<SID>/HDB00/<hananode> we use a relative path to cwd this gives us a <sid>adm permitted directory
-                    #     however we go one level up (..) to have the file accessable for all SAP HANA swarm nodes
-                    #
-                    fallbackFileObject = open("../.crm_attribute.stage.{0}".format( mySite ), "w")
-                    if globHOOK:
-                        fallbackFileObject.write("hana_{0}_glob_srHook = {1}".format(mysid, mySRS))
-                    fallbackFileObject.write("hana_{0}_site_srHook_{1} = {2}".format(mysid, mySite, mySRS))
-                    fallbackFileObject.close()
-                    #
-                    # release the stage file to the original name (move is used to be atomic)
-                    #      .crm_attribute.stage.<site> is renamed to .crm_attribute.<site>
-                    #
-                    os.rename( "../.crm_attribute.stage.{0}".format( mySite ), "../.crm_attribute.{0}".format( mySite ) )
-            return 0
-except NameError as e:
-        print("Could not find base class ({0})".format(e))
+                mySRS = "SFAIL"
+        if (mySRS == ""):
+            self.tracer.info("SAPHanaSR (%s) 001" % (self.__class__.__name__))
+            myMSG = "### Ignoring bad SR status because of is_in_sync=True ###"
+            self.tracer.info("SAPHanaSR (%s) 002" % (self.__class__.__name__))
+        else:
+            myCMD = "sudo /usr/sbin/crm_attribute -n hana_%s_glob_srHook -v %s -t crm_config -s SAPHanaSR" % (mysid, mySRS)
+            rc = os.system(myCMD)
+            myMSG = "CALLING CRM: <" + myCMD + "> rc=" + str(rc)
+        self.tracer.info("SAPHanaSR %s.srConnectionChanged method called with Dict=%s ###\n" % (self.__class__.__name__, ParamDict))
+        self.tracer.info("SAPHanaSR %s \n" % (myMSG))
+        return 0
