+++ SAPHana/srHook/global.ini
@@ -4,12 +4,11 @@ provider = SAPHanaSR
 path = /usr/share/SAPHanaSR-ScaleOut
 execution_order = 1
 
-[ha_dr_provider_SAPHanaSrTakeoverBlocker]
-provider = SAPHanaSrTakeoverBlocker
-path = /usr/share/SAPHanaSR-ScaleOut/
+[ha_dr_provider_SAPHanaSrMultiTarget]
+provider = SAPHanaSrMultiTarget
+path = /usr/share/SAPHanaSR-ScaleOut
 execution_order = 2
 
 [trace]
 ha_dr_saphanasr = info
-ha_dr_saphanasrtakeoverblocker = info
-
+ha_dr_saphanasrmultitarget = info
